Empty pipe.py & response.py into layer.py
